+++ src/components/Banner/Banner.stories.tsx
@@ -1,6 +1,5 @@
 import { VStack } from "@chakra-ui/react";
 import type { Meta, StoryObj } from "@storybook/react";
-
 import Banner from "./Banner";
 import Button from "../Button/Button";
 import ButtonGroup from "../ButtonGroup/ButtonGroup";
@@ -32,7 +31,7 @@ const meta: Meta<typeof Banner> = {
     icon: { control: false },
     id: { control: false },
     isDismissible: { control: { type: "boolean" } },
-    type: {
+    variant: {
       control: { type: "select" },
       options: messageVariantsArray,
       table: { defaultValue: { summary: "neutral" } },
@@ -109,7 +108,7 @@ export const TypeVariants: Story = {
               faucibus dolor auctor.
               `}
               heading="Informative Banner"
-              type="informative"
+              variant="informative"
             />
           </>,
           "Communicates helpful information or an important attribute.",
@@ -126,7 +125,7 @@ export const TypeVariants: Story = {
               faucibus dolor auctor.
               `}
               heading="Positive Banner"
-              type="positive"
+              variant="positive"
             />
           </>,
           "Indicates a constructive or successful state.",
@@ -146,7 +145,7 @@ export const TypeVariants: Story = {
                 </>
               }
               heading="Negative Banner"
-              type="negative"
+              variant="negative"
             />
           </>,
           "Informs users of problems or errors that require potential action to correct.",
@@ -163,7 +162,7 @@ export const TypeVariants: Story = {
               faucibus dolor auctor.
               `}
               heading="Warning Banner"
-              type="warning"
+              variant="warning"
             />
           </>,
           "Communicates cautionary or time-sensitive information.",
@@ -180,7 +179,7 @@ export const TypeVariants: Story = {
               faucibus dolor auctor.
               `}
               heading="Recommendation Banner"
-              type="recommendation"
+              variant="recommendation"
             />
           </>,
           "Highlights a suggestion that will improve the experience and achieve better results.",
@@ -216,7 +215,7 @@ export const BannerHeading: Story = {
           </>
         }
         heading={<Heading level="h5">Custom H5 Heading</Heading>}
-        type="positive"
+        variant="positive"
       />
       <Banner
         content={
@@ -228,7 +227,7 @@ export const BannerHeading: Story = {
             sagittis lacus vel augue laoreet rutrum faucibus dolor auctor.
           </>
         }
-        type="warning"
+        variant="warning"
       />
     </VStack>
   ),
@@ -255,7 +254,7 @@ export const CustomBannerIcon: Story = {
             marginTop="xxxs"
           />
         }
-        type="informative"
+        variant="informative"
       />
       <Banner
         content={
@@ -275,7 +274,7 @@ export const CustomBannerIcon: Story = {
             marginTop="xxxs"
           />
         }
-        type="positive"
+        variant="positive"
       />
       <Banner
         content={
@@ -295,7 +294,7 @@ export const CustomBannerIcon: Story = {
             marginTop="xxxs"
           />
         }
-        type="negative"
+        variant="negative"
       />
       <Banner
         content={
@@ -350,7 +349,7 @@ export const HTMLContent: Story = {
               faucibus mollis interdum.
             </Text>
             <ButtonGroup>
-              <Button buttonType="secondary" id="group-1">
+              <Button variant="secondary" id="group-1">
                 Button
               </Button>
               <Button id="group-2">Submit</Button>
@@ -389,7 +388,7 @@ export const Dismissible: Story = {
       }
       heading="Dismissible Banner"
       isDismissible
-      type="neutral"
+      variant="neutral"
     />
   ),
 };
@@ -400,7 +399,7 @@ export const StringContentWithHTML: Story = {
         "<p>Cras mattis consectetur purus sit amet fermentum. Maecenas faucibus mollis interdum.</p><p>Morbi leo risus, porta ac consectetur ac, vestibulum at eros. <b>Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus</b>. <a href='#'>This is a link</a>.</p>"
       }
       heading="String content with HTML"
-      type="neutral"
+      variant="neutral"
     />
   ),
 };