+++ src/components/Button/buttonChangelogData.ts
@@ -19,6 +19,7 @@ export const changelogData: ChangelogData[] = [
       "Changes theme name from `CustomButton` to `ReservoirButton` for consistency.",
       "Removes explicit `className` prop as interface can be extended to include Chakra prop or HTML attribute types.",
       "Removes `type` and `onClick` props (as HTML equivalents may be used) and updates docs.",
+      "Renames `buttonType` to `variant`.",
       "Replaces `buttonSizesArray` with shared sizes array.",
     ],
   },