+++ src/components/Notification/Notification.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 
 import Icon from "../Icons/Icon";
 import Link from "../Link/Link";
-import Notification, { notificationTypesArray } from "./Notification";
+import Notification, { notificationVariantsArray } from "./Notification";
 import Text from "../Text/Text";
 import Heading from "../Heading/Heading";
 
@@ -19,11 +19,11 @@ const meta: Meta<typeof Notification> = {
       table: { defaultValue: { summary: "false" } },
     },
     notificationContent: { control: false },
-    notificationType: {
+    variant: {
       control: {
         type: "radio",
       },
-      options: notificationTypesArray,
+      options: notificationVariantsArray,
       table: { defaultValue: { summary: "standard" } },
     },
   },
@@ -53,7 +53,7 @@ export const WithControls: Story = {
         luctus, nisi erat porttitor ligula.
       </>
     ),
-    notificationType: "standard",
+    variant: "standard",
     showIcon: true,
   },
   parameters: {
@@ -88,7 +88,7 @@ export const NotificationHeading: Story = {
         }
       />
       <Notification
-        notificationType="announcement"
+        variant="announcement"
         notificationHeading="Announcement Notification"
         notificationContent={
           <>
@@ -102,7 +102,7 @@ export const NotificationHeading: Story = {
         }
       />
       <Notification
-        notificationType="warning"
+        variant="warning"
         notificationHeading="Warning Notification"
         notificationContent={
           <>
@@ -128,7 +128,7 @@ export const NotificationHeading: Story = {
         }
       />
       <Notification
-        notificationType="announcement"
+        variant="announcement"
         notificationContent={
           <>
             This is an "announcement" Notification without a heading. Cras
@@ -141,7 +141,7 @@ export const NotificationHeading: Story = {
         }
       />
       <Notification
-        notificationType="warning"
+        variant="warning"
         notificationContent={
           <>
             This is a "warning" Notification without a heading. Cras mattis
@@ -293,7 +293,7 @@ export const Dismissible: Story = {
           nascetur ridiculus mus.
         </>
       }
-      notificationType="announcement"
+      variant="announcement"
     />
   ),
 };